+++ packages/terra-responsive-element/src/ResponsiveElement.jsx
@@ -4,6 +4,11 @@ import ResizeObserver from 'resize-observer-polyfill';
 import 'terra-base/lib/baseStyles';
 import breakpoints from './breakpoints.module.scss';
 
+const DependentViewport = {
+  WINDOW: 'window',
+  PARENT: 'parent',
+};
+
 /* Disabled for the defaultElement which is used implicitly */
 /* eslint-disable react/no-unused-prop-types */
 const propTypes = {
@@ -122,3 +127,4 @@ ResponsiveElement.propTypes = propTypes;
 ResponsiveElement.defaultProps = defaultProps;
 
 export default ResponsiveElement;
+export { DependentViewport };